feat(metric_alerts): Add logs for session and metric subscriptions for comparison.
We want to switch over from using sessions to power crash rate alerts to metrics. Before we do this,
we want to validate the metrics are producing the same or similar values to sessions.

We'll log this to GCP and then download the log files and compare locally. When creating test
subscriptions for metrics based on existing session subscriptions I'll log a mapping that can be
used locally for comparison.

For logging metric subscriptions I'm just reusing an old handler that we used before. I'll make
sure the old test subscriptions are removed first before we merge this.